Show total rewards for ended programs
