+++ src/components/Pool/PoolControls/Claim.tsx
@@ -1,30 +1,20 @@
 import React, { useCallback, useMemo } from 'react'
 // @ts-ignore
-import TokenAmount from 'token-amount'
 import { useClaimRewards } from '../../../hooks/useContract'
 import { useAccountModule } from '../../Account/AccountModuleProvider'
-import AmountCard from '../../AmountCard/AmountCard'
 import { usePoolBalance } from '../PoolBalanceProvider'
 import { usePoolInfo } from '../PoolInfoProvider'
 import { ValidationStatus } from '../types'
 import ControlButton from './ControlButton'
+import TotalRewardsCard from './TotalRewardsCard'
 
 function Claim(): JSX.Element {
-  const { rewardToken, contractGroup } = usePoolInfo()
-  const { rewardsBalanceInfo, formattedDigits } = usePoolBalance()
+  const { contractGroup } = usePoolInfo()
+  const { rewardsBalanceInfo } = usePoolBalance()
   const { showAccount } = useAccountModule()
   const handleClaim = useClaimRewards(contractGroup)
 
-  const [rewardsBalance, rewardsBalanceStatus] = rewardsBalanceInfo
-
-  const formattedRewardsBalance = useMemo(
-    (): string | null =>
-      rewardsBalance &&
-      new TokenAmount(rewardsBalance, rewardToken.decimals).format({
-        digits: formattedDigits,
-      }),
-    [rewardsBalance, rewardToken.decimals, formattedDigits]
-  )
+  const [rewardsBalance] = rewardsBalanceInfo
 
   const validationStatus = useMemo((): ValidationStatus => {
     if (!rewardsBalance) {
@@ -55,16 +45,7 @@ function Claim(): JSX.Element {
 
   return (
     <form onSubmit={handleSubmit}>
-      <AmountCard
-        label="Total rewards generated"
-        tokenGraphic={rewardToken.graphic}
-        suffix={rewardToken.symbol}
-        value={formattedRewardsBalance ? formattedRewardsBalance : '0'}
-        loading={rewardsBalanceStatus === 'loading'}
-        css={`
-          margin-top: 10px;
-        `}
-      />
+      <TotalRewardsCard />
       <ControlButton
         status={validationStatus}
         labels={{
