+++ src/components/Pool/PoolControls/Withdraw.tsx
@@ -11,6 +11,7 @@ import AmountInput from '../../AmountInput/AmountInput'
 import { usePoolBalance } from '../PoolBalanceProvider'
 import { usePoolInfo } from '../PoolInfoProvider'
 import ControlButton from './ControlButton'
+import TotalRewardsCard from './TotalRewardsCard'
 import useInputValidation from './useInputValidation'
 
 type WithdrawProps = {
@@ -127,12 +128,19 @@ function Withdraw({ exitAllBalance }: WithdrawProps): JSX.Element {
       )}
 
       <AmountCard
-        label={`Amount available to withdraw`}
+        label={
+          exitAllBalance
+            ? 'Total amount staked'
+            : 'Amount available to withdraw'
+        }
         tokenGraphic={stakeToken.graphic}
         suffix={stakeToken.symbol}
         value={formattedStakedBalance ? formattedStakedBalance : '0'}
         loading={stakedBalanceStatus === 'loading'}
       />
+
+      {exitAllBalance && <TotalRewardsCard />}
+
       <ControlButton
         status={filteredValidationStatus}
         labels={{
@@ -141,7 +149,7 @@ function Withdraw({ exitAllBalance }: WithdrawProps): JSX.Element {
             ? 'You have no funds to withdraw'
             : 'Insufficient stake balance',
           noAmount: 'Enter an amount',
-          valid: 'Withdraw',
+          valid: exitAllBalance ? 'Withdraw all' : 'Withdraw',
           loading: 'Loading…',
         }}
       />
